SQL-to-Modern-Dashboards/requirements.txt
streamlit
google-generativeai
python-dotenv
plotly